checkpoint: moves "/URLComponent/Origin/" to "/URLComponent/DepOrigin/"
