+++ src/library/URLComponent/DepOrigin/ParsingError.ts
