+++ src/library/URLComponent/DepOrigin/definition.declared.augmentation.ts
