+++ src/library/URLComponent/DepOrigin/definition.declared.ts
