+++ src/library/URLComponent/DepOrigin/definition.declared.withAugmentation.ts
