+++ src/library/URLComponent/DepOrigin/exports.object.primary.ts
