+++ src/library/URLComponent/DepOrigin/index.ts
